+++ History.md
@@ -79,6 +79,7 @@ Unreleased
 
 * added; `BankAccountAuthorizedAt` to `Subscription`
 * added; `IpAddress` to `Transaction`
+* added; CreditCardType.DinersClub
 
 1.2.0 (stable) / 2015-04-28
 ==================